Rename fitters.linear to fitters.least_squares
