+++ src/prfmodel/fitters/__init__.py
@@ -1,7 +1,7 @@
 """Model fitters."""
 
 from .grid import GridFitter
-from .linear import LeastSquaresFitter
+from .least_squares import LeastSquaresFitter
 from .sgd import SGDFitter
 
 __all__ = [
